Generación del mapa y sus elementos
Se ha implementado un sistema de generación de mapas de manera procedural, así como la colocación aleatoria de edificios, cajas, decoración y enemigos sobre él de manera aleatoria, así como un sistema de daño de manera que el mapa va reduciendo su área segura para acelerar la interacción entre las entidades del juego

+++ TheOnlyOne/Assets/Scripts/ScriptableObjectsGenerator/WeaponBlueprint.cs
@@ -23,8 +23,8 @@ public class WeaponBlueprint : ScriptableObject
     public int damage;
     [Space]
     [Header("Sway")]
-    public float swayIntensity;
-    public float swaySpeed;
+    public int swayIntensity;
+    public int swaySpeed;
     [Space]
     [Header("ShootRecoil")]
     public Vector3 recoilRotation;
